requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cloudinary==1.28.0
dj-database-url==0.5.0
Django==3.2.9
django-bootstrap3==21.1
django-braces==1.15.0
django-filter==21.1
django-heroku==0.3.1
django-model-utils==4.2.0
django-star-ratings==0.9.2
djangorestframework==3.12.4
gunicorn==20.1.0
importlib-metadata==4.8.2
Markdown==3.3.6
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
